mm_to_json/mm_to_json_py/pyproject.toml
[project]
name = "mm-to-json-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "access-parser>=0.0.6",
    "pandas>=2.3.3",
    "JPype1>=1.5.0",
    "JayDeBeApi>=1.2.3",
    "ruff>=0.9.0",
    "reportlab>=4.2.5",
]
